New: db.all({ sql, nestTables: true }), fixes #443
This commit changes the Statement callback footprint from `function(err,
result)` to `function(err, result, fields)`. The default result
structure is changed from `{ field1: value1, field2: value2 }` to
`[ value1, value2 ]` to be compliant with the fields array.

The newly added fields is in such format:

```js
[ { name: 'field1', table: 'table1' },
  { name: 'field2', table: 'table1' } ]
```

In this way, when a JOIN query like `SELECT foo.*, bar.* ...` happens,
the values in result won't be collided with each other. And the result
can be properly processed with the help of fields.

To be backward compliant, `Database.prototype.[get|all]` are overriden.
The default `.all(sql, callback)` still works the same, with the result
containing objects. The added method type is `.all({ sql, nestTables })` in
which the `nestTables` is default to false.

If `nestTables` is true, the returned `rows` is an array of nested objects
in following structure:

```js
[ { table1: { field1: value1, field2: value2 },
    table2: { ... } },
  { table1: { ... },
    table2: { ... } } ]
```